Authentication and Authorization Added

+++ backend/controllers/controller.js
@@ -2,13 +2,14 @@ const asyncHandler = require('express-async-handler')
 
 
 const Goal = require("../models/models")
+const User = require("../models/userModel")
 // @Description : GET
 // @Routing : /api/goals/
 // @Access : Private
 
 const getGoals = asyncHandler(async (req, res) => {
 
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
 
     res.status(200).json(goals)
 
@@ -28,7 +29,8 @@ const setGoals = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id,
     })
 
 
@@ -46,6 +48,20 @@ const putGoal = asyncHandler(async (req, res) => {
     if(!goal){
         res.status(400)
         throw new Error('Goal Not Found!')
+    }
+
+    // Check Credentials
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(401)
+        throw new Error("User Not Found")
+    }
+
+    // Logged In User Matches the Goal User
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error("User Not Authorized")
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true,})
@@ -77,7 +93,22 @@ const deleteGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal Not Found!')
     }
 
+    // Check Credentials
+    const user = await User.findById(req.user.id)
+
+    if(!user){
+        res.status(401)
+        throw new Error("User Not Found")
+    }
+
+    // Logged In User Matches the Goal User
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error("User Not Authorized")
+    }
+
     await Goal.deleteOne()
+    // await Goal.remove()
 
 
     res.status(200).json({ id: req.params.id })
